requirements.txt
torch>=2.0.0
transformers>=4.30.0
peft>=0.5.0
datasets>=2.12.0
wandb>=0.15.0
tqdm>=4.65.0
numpy>=1.24.0
scikit-learn>=1.2.0
huggingface-hub>=0.16.0
python-dotenv>=0.19.0
openai>=1.0.0
accelerate>=0.20.0